// src/oem/main/oem-pricing-models/oem-pricing-model.enums/unit-duration.enum.ts
export enum UnitDurationEnum {
  CONSUMED = 'Consumed',
  CONSUMED_PER_CALENDAR_YEAR = 'Consumed Per Calendar Year',
  CONSUMED_PER_CALENDAR_MONTH = 'Consumed Per Calendar Month',
  CONSUMED_PER_DAY = 'Consumed Per Day',
  CONSUMED_PER_WEEK = 'Consumed Per Week',
  CONSUMED_PER_30 = 'Consumed Per 30 Days',
  CONSUMED_PER_31 = 'Consumed Per 31 Days',
  CONSUMED_PER_365 = 'Consumed Per 365 Days',
  PER_ONE_TIME = 'One-Time / Non-Recurring',
  PER_CALENDAR_YEAR = 'Per Calendar Year',
  PER_CALENDAR_MONTH = 'Per Calendar Month',
  PER_SECOND = 'Per Seconds',
  PER_HOUR = 'Per Hour',
  PER_DAY = 'Per Day',
  PER_WEEK = 'Per Week',
  PER_30 = 'Per 30 Days',
  PER_31 = 'Per 31 Days',
  PER_365 = 'Per 365 Days',
}